Multi-threaded Fibonnaci Calculator Attempt.
I found something really interesting with this:
My implementation should run two parallel fib functions in threads,
accessing the same fib_memo vector. Only one of the threads actually
calculates anything, and it does it for both fib-calls. The best part
is, that the thread with the smaller fib-number to be calculated, does
the full computation, and not the thread with the larger one.

An explanation of this can be found in future-commits in the docs of
this crate!